+++ doc/release-notes.md
@@ -34,6 +34,19 @@ Notable Changes
 
 (Developers: add your notes here as part of your pull requests whenever possible)
 
+Low-level RPC changes
+---------------------
+
+- The new database model no longer stores information about transaction
+  versions of unspent outputs. This means that:
+  - The `gettxout` RPC no longer has a `version` field in the response.
+  - The `gettxoutsetinfo` RPC reports `hash_serialized_2` instead of `hash_serialized`,
+    which does not commit to the transaction versions of unspent outputs, but does
+    commit to the height and coinbase/coinstake information.
+  - The `getutxos` REST path no longer reports the `txvers` field in JSON format,
+    and always reports 0 for transaction versions in the binary format
+
+
 *version* Change log
 ==============